Renaming: 'group_processing' -> 'stars_group'
