--- alcor/services/group_processing/__init__.py
@@ -1,1 +1,0 @@
-from .service import process_stars_group
