+++ alcor/services/stars_group/__init__.py
@@ -1,0 +1,1 @@
+from .service import process
