+++ alcor/services/stars_group/elimination.py
