+++ alcor/services/stars_group/utils.py
